+++ src/main/java/com/culturestamp/back/CultureStampBackApplication.java
@@ -1,13 +1,13 @@
-package com.culturestampserver;
+package com.culturestamp.back;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class CultureStampServerApplication {
+public class CultureStampBackApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(CultureStampServerApplication.class, args);
+		SpringApplication.run(CultureStampBackApplication.class, args);
 	}
 
 }
