+++ src/test/java/com/culturestamp/back/CultureStampServerApplicationTests.java
@@ -1,4 +1,4 @@
-package com.culturestampserver;
+package com.culturestamp.back;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
